+++ cluster.go
@@ -11,7 +11,7 @@ import (
 
 	"google.golang.org/grpc/connectivity"
 
-	"github.com/yandex-cloud/ydb-go-sdk/v2/timeutil"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/timeutil"
 )
 
 const (