+++ example/basic_example_v1/series.go
@@ -1,10 +1,6 @@
 package main
 
 import (
-	"github.com/yandex-cloud/ydb-go-sdk/v2"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/connect"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/example/internal/cli"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/table"
 	"bytes"
 	"context"
 	"flag"
@@ -13,6 +9,11 @@ import (
 	"path"
 	"text/template"
 	"time"
+
+	"github.com/YandexDatabase/ydb-go-sdk/v2"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/connect"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/example/internal/cli"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/table"
 )
 
 type templateConfig struct {