+++ example/containers/containers.go
@@ -1,16 +1,16 @@
 package main
 
 import (
-	"github.com/yandex-cloud/ydb-go-sdk/v2/connect"
 	"bytes"
 	"context"
 	"flag"
 	"fmt"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/connect"
 	"log"
 	"text/template"
 
-	"github.com/yandex-cloud/ydb-go-sdk/v2/example/internal/cli"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/table"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/example/internal/cli"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/table"
 )
 
 var query = template.Must(template.New("fill database").Parse(`