+++ example/secondary_index_builtin/describe.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"path"
 
-	"github.com/yandex-cloud/ydb-go-sdk/v2/table"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/table"
 )
 
 func doDescribe(