+++ experimental/experimental.go
@@ -3,11 +3,11 @@ package experimental
 import (
 	"bytes"
 	"context"
+	"github.com/YandexDatabase/ydb-go-sdk/v2"
 
-	"github.com/yandex-cloud/ydb-go-sdk/v2"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/api/protos/Ydb"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/api/protos/Ydb_Experimental"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/internal"
+	"github.com/YandexDatabase/ydb-go-genproto/protos/Ydb"
+	"github.com/YandexDatabase/ydb-go-genproto/protos/Ydb_Experimental"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/internal"
 )
 
 type Client struct {