+++ value.go
@@ -1,7 +1,7 @@
 package ydb
 
 import (
-	"github.com/yandex-cloud/ydb-go-sdk/v2/internal"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/internal"
 )
 
 type Value interface {