+++ ydbsql/connector.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"database/sql/driver"
 	"fmt"
+	"github.com/YandexDatabase/ydb-go-sdk/v2"
 	"sync"
 	"time"
 
-	"github.com/yandex-cloud/ydb-go-sdk/v2"
-	"github.com/yandex-cloud/ydb-go-sdk/v2/table"
+	"github.com/YandexDatabase/ydb-go-sdk/v2/table"
 )
 
 var (